+++ changelog.markdown
@@ -102,3 +102,5 @@
 * Tested with **ghc-8.0.1**.
 ## 0.9.0.15
 * Added default RTS-options to **regexchar.cabal**.
+* Removed dependency on **Distribution.Package.PackageName**
+